feat: 이메일이 저장소에 중복 시 예외 발생 추가
