+++ src/test/java/roomescape/domain/MemberNameTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain.member;
+package roomescape.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -6,16 +6,17 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import roomescape.domain.member.MemberName;
 
-class MemberPasswordTest {
+class MemberNameTest {
 
     @ParameterizedTest
     @NullAndEmptySource
-    @ValueSource(strings = {" ", "   ", "     "})
-    @DisplayName("비밀번호가 비어있거나 공백이면 예외를 발생한다.")
-    void validateMemberPassword(String given) {
+    @ValueSource(strings = {" ", "  ", "   "})
+    @DisplayName("사용자 이름이 공백이면 예외를 발생한다.")
+    void validateMemberName(String given) {
         //when //then
-        assertThatThrownBy(() -> new MemberPassword(given))
+        assertThatThrownBy(() -> new MemberName(given))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
