+++ src/test/java/roomescape/domain/MemberPasswordTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain.member;
+package roomescape.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -6,16 +6,17 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import roomescape.domain.member.MemberPassword;
 
-class MemberEmailTest {
+class MemberPasswordTest {
 
     @ParameterizedTest
     @NullAndEmptySource
-    @ValueSource(strings = {" ", "abr", "12341231"})
-    @DisplayName("이메일 형식이 아닌 입력값은 예외가 발생한다.")
-    void validateEmail(String input) {
+    @ValueSource(strings = {" ", "   ", "     "})
+    @DisplayName("비밀번호가 비어있거나 공백이면 예외를 발생한다.")
+    void validateMemberPassword(String given) {
         //when //then
-        assertThatThrownBy(() -> new MemberEmail(input))
+        assertThatThrownBy(() -> new MemberPassword(given))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
